==> 02_rename_gtfs_stations_to_wmata_names.ipynb <==
# Import packages

import pandas as pd
import geopandas as gpd

from helper_functions import open_processed_gtfs, open_wmata_station_data

# Notebook settings

GTFS_FOLDER = 'GTFS_20190914'

# Load in Data

# WMATA stations
wmata_stations = open_wmata_station_data(r'Data\GIS_RAIL_STN_PNT.csv')
wmata_stations.head(3)

# Processed GTFS stations
gtfs_stations = open_processed_gtfs(f"Intermediates\\{GTFS_FOLDER}\\gtfs_stations.geojson")
print(len(gtfs_stations))
gtfs_stations.head(3)

# Find WMATA station for each GTFS station and correct station names

# Set both gdf's to a local XY crs
wmata_stations = wmata_stations.to_crs('EPSG:2248')
gtfs_stations = gtfs_stations.to_crs('EPSG:2248')

corrected_gtfs_stations = gtfs_stations.copy()
# Save correction to save time correcting other files
correction_dict = {}
# First find matching stations and correct stop_id and stop_name
for index, stop in gtfs_stations.iterrows():
    wmata_stations['distance'] = wmata_stations.distance(stop['geometry'])
    wmata_stations = wmata_stations.sort_values(by='distance')
    
    matching_station = wmata_stations['STATION_ID'].values[0]
    corrected_gtfs_stations.at[index, 'stop_id'] = matching_station
    correction_dict[stop['stop_id']] = matching_station

# Then correct from-to stop lists
for index, stop in corrected_gtfs_stations.iterrows():

    connections = []
    for connection in stop['connections']:
        connection_1 = connection[0]
        connection_2 = connection[1]
        if connection_1 != 'None':
            connection_1 = correction_dict[connection_1]
        if connection_2 != 'None':
            connection_2 = correction_dict[connection_2]

        connections.append([connection_1, connection_2])
    corrected_gtfs_stations.at[index, 'connections'] = connections

corrected_gtfs_stations = corrected_gtfs_stations.to_crs('EPSG:4326')  # All gdf are saved in geodesic, so we convert back
corrected_gtfs_stations.to_file(f"Intermediates\\{GTFS_FOLDER}\\gtfs_stations_corrected.geojson", driver="GeoJSON")
corrected_gtfs_stations.head(5)

# Correct links

gtfs_shapes = gpd.read_file(f"Intermediates\\{GTFS_FOLDER}\\gtfs_links.geojson")

gtfs_shapes_corrected = gtfs_shapes.copy()
for index, shape in gtfs_shapes.iterrows():
    gtfs_shapes_corrected.at[index, 'from_stop'] = correction_dict[shape['from_stop']]
    gtfs_shapes_corrected.at[index, 'to_stop'] = correction_dict[shape['to_stop']]
    gtfs_shapes_corrected.at[index, 'link_id'] = f"{gtfs_shapes_corrected.at[index, 'from_stop']}_to_{gtfs_shapes_corrected.at[index, 'to_stop']}"

gtfs_shapes_corrected.to_file(f"Intermediates\\{GTFS_FOLDER}\\gtfs_links_corrected.geojson", driver="GeoJSON", index=False)
gtfs_shapes_corrected.head(3)

# Correct stop_datetimes

gtfs_stoptimes = gpd.read_file(f"Intermediates\\{GTFS_FOLDER}\\gtfs_stop_datetimes.csv", index=False)
gtfs_stoptimes.head(3)

correction_dict_stings = {str(key): value for key, value in correction_dict.items()}

gtfs_stoptimes_corrected = gtfs_stoptimes.copy()
gtfs_stoptimes_corrected['stop'] = gtfs_stoptimes_corrected['stop'].replace(correction_dict_stings)
gtfs_stoptimes_corrected['previous_stop'] = gtfs_stoptimes_corrected['previous_stop'].replace(correction_dict_stings)
gtfs_stoptimes_corrected['next_stop'] = gtfs_stoptimes_corrected['next_stop'].replace(correction_dict_stings)

gtfs_stoptimes_corrected.to_csv(f"Intermediates\\{GTFS_FOLDER}\\gtfs_stop_datetimes_corrected.csv", index=False)
gtfs_stoptimes_corrected.head(3)

print(f"{len(gtfs_stoptimes_corrected):_}")
# Get first and last dates to filter wmata stop datetimes by in next notebook
print(gtfs_stoptimes_corrected['arrival_time'].min(), gtfs_stoptimes_corrected['arrival_time'].max())
