==> 03_prepare_wmata_data.ipynb <==
# Import packages

import pandas as pd
import geopandas as gpd

from helper_functions import open_wmata_station_data

# Notebook settings

FILES_TO_USE = ['Train_Movement_20190801_20191231.csv',
                'Train_Movement_20200101_20201231.csv']
WMATA_STATION_FILEPATH = r'Data\GIS_RAIL_STN_PNT.csv'
# Dates gotten from last codeblock of previous notebook, follows from GTFS data
BEGIN_DATE = '2019-09-14'
END_DATE = '2020-03-12'

# Open Data

wmata_stop_datetimes_dfs = []

for file in FILES_TO_USE:
    wmata_stop_datetimes_i = pd.read_csv(f"Data\\trainmovement\\{file}", usecols=['TRIP_ROUTE', 'BERTH_ID', 'NEXT_BERTH', 'ARRIVAL', 'TRIP_KEY'])
    wmata_stop_datetimes_dfs.append(wmata_stop_datetimes_i)

wmata_stop_datetimes = pd.concat(wmata_stop_datetimes_dfs)
print(f"{len(wmata_stop_datetimes):_}")
wmata_stop_datetimes.head(3)

wmata_stations = open_wmata_station_data(WMATA_STATION_FILEPATH)
wmata_stations['STATION_ID_LIST'] = [station_id.split('/') for station_id in wmata_stations['STATION_ID'].values]
print(len(wmata_stations))

# Filter to same date range as GTFS

# Convert arrival times to useable datetime objects
wmata_stop_datetimes['ARRIVAL'] = pd.to_datetime(wmata_stop_datetimes['ARRIVAL'], format='%d-%b-%Y %H:%M:%S')
wmata_stop_datetimes = wmata_stop_datetimes.sort_values(by=['ARRIVAL'])
wmata_stop_datetimes = wmata_stop_datetimes[wmata_stop_datetimes['ARRIVAL'] >= BEGIN_DATE]
wmata_stop_datetimes = wmata_stop_datetimes[wmata_stop_datetimes['ARRIVAL'] < END_DATE]
print(f"{len(wmata_stop_datetimes):_}")
wmata_stop_datetimes.head(3)

# Convert data to useable format

# Rename line names to full names for better readability
wmata_stop_datetimes['TRIP_ROUTE'] = wmata_stop_datetimes['TRIP_ROUTE'].replace({'RD': 'RED',
                                                                                 'SV': 'SILVER',
                                                                                 'OR': 'ORANGE',
                                                                                 'BL': 'BLUE',
                                                                                 'GR': 'GREEN',
                                                                                 'YL': 'YELLOW',
                                                                                 'BL|OR|SV': 'BLUE, ORANGE, SILVER',
                                                                                 'BL|SV': 'BLUE, SILVER',
                                                                                 'BL|YL': 'BLUE, YELLOW',
                                                                                 'GR|YL': 'GREEN, YELLOW',
                                                                                 'OR|SV': 'ORANGE, SILVER'})
# Convert berth_id's to station/stop ids by removing berth-specific affix
wmata_stop_datetimes['BERTH_ID'] = [value[:3] if pd.notna(value) else pd.NA for value in wmata_stop_datetimes['BERTH_ID'].values]
wmata_stop_datetimes['NEXT_BERTH'] = [value[:3] if pd.notna(value) else pd.NA for value in wmata_stop_datetimes['NEXT_BERTH'].values]

wmata_stop_datetimes.head(3)

# Correct stations names

# Some stations have double names (B01 and F01 are the same phyisical station for example), and so we create a correction dictionary
correction_dict = {}
for berth_id in wmata_stop_datetimes['BERTH_ID'].unique():
    if pd.notna(berth_id):
        matches = wmata_stations[[berth_id in station_list for station_list in wmata_stations['STATION_ID_LIST'].values]]
        if len(matches) > 0:
            correction_dict[berth_id] = matches['STATION_ID'].values[0]

# And then correct all stop names in the datetimes dataset
wmata_stop_datetimes['BERTH_ID'] = wmata_stop_datetimes['BERTH_ID'].replace(correction_dict)
wmata_stop_datetimes['NEXT_BERTH'] = wmata_stop_datetimes['NEXT_BERTH'].replace(correction_dict)
wmata_stop_datetimes = wmata_stop_datetimes.rename(columns={'BERTH_ID': 'STOP_ID', 'NEXT_BERTH': 'NEXT_STOP_ID'})
wmata_stop_datetimes.head(5)

# Save file

wmata_stop_datetimes.to_csv(f"Intermediates\\wmata_stop_datetimes.csv", index=False)

### Open file

if wmata_dataset_year == 2019:
    path = f"Data\\trainmovement\\Train_Movement_{wmata_dataset_year}0801_{wmata_dataset_year}1231.csv"
else:
    path = f"Data\\trainmovement\\Train_Movement_{wmata_dataset_year}0101_{wmata_dataset_year}1231.csv"

wmata_stoptimes = pd.read_csv(path, usecols=['TRIP_ROUTE', 'BERTH_ID', 'NEXT_BERTH', 'ARRIVAL', 'TRIP_KEY'])
wmata_stoptimes.head(3)

### Convert column values to desired formats

wmata_stoptimes['TRIP_ROUTE'] = wmata_stoptimes['TRIP_ROUTE'].replace({'RD': 'RED',
                                                                       'SV': 'SILVER',
                                                                       'OR': 'ORANGE',
                                                                       'BL': 'BLUE',
                                                                       'GR': 'GREEN',
                                                                       'YL': 'YELLOW',
                                                                       'BL|OR|SV': 'BLUE, ORANGE, SILVER',
                                                                       'BL|SV': 'BLUE, SILVER',
                                                                       'BL|YL': 'BLUE, YELLOW',
                                                                       'GR|YL': 'GREEN, YELLOW',
                                                                       'OR|SV': 'ORANGE, SILVER'})
wmata_stoptimes['BERTH_ID'] = [value[:3] if pd.notna(value) else pd.NA for value in wmata_stoptimes['BERTH_ID'].values]
wmata_stoptimes['NEXT_BERTH'] = [value[:3] if pd.notna(value) else pd.NA for value in wmata_stoptimes['NEXT_BERTH'].values]
wmata_stoptimes['ARRIVAL'] = pd.to_datetime(wmata_stoptimes['ARRIVAL'], format='%d-%b-%Y %H:%M:%S')
wmata_stoptimes = wmata_stoptimes.sort_values(by=['ARRIVAL'])
wmata_stoptimes

# WMATA stations
wmata_stations = pd.read_csv(r'Data\GIS_RAIL_STN_PNT.csv', sep=',', usecols=['NAME', 'X_COORD', 'Y_COORD', 'DISPLAY_NAME'])
wmata_stations = gpd.GeoDataFrame(wmata_stations, geometry=gpd.points_from_xy(wmata_stations['X_COORD'], wmata_stations['Y_COORD']))
wmata_stations['STATION_ID'] = [name.split('(')[-1][:-1] for name in wmata_stations['DISPLAY_NAME']]
wmata_stations = wmata_stations.drop(columns=['X_COORD', 'Y_COORD', 'DISPLAY_NAME'])
wmata_stations = wmata_stations.set_crs('EPSG:4326')
wmata_stations = wmata_stations.to_crs('EPSG:2248')
wmata_stations.head(3)

wmata_stations = pd.read_csv(r'Data\GIS_RAIL_STN_PNT.csv', sep=',', usecols=['NAME', 'X_COORD', 'Y_COORD', 'DISPLAY_NAME'])
wmata_stations = gpd.GeoDataFrame(wmata_stations, geometry=gpd.points_from_xy(wmata_stations['X_COORD'], wmata_stations['Y_COORD']))
wmata_stations['STATION_ID'] = [name.split('(')[-1][:-1] for name in wmata_stations['DISPLAY_NAME']]
wmata_stations['STATION_ID_LIST'] = [station_id.split('/') for station_id in wmata_stations['STATION_ID'].values]
wmata_stations = wmata_stations.drop(columns=['X_COORD', 'Y_COORD', 'DISPLAY_NAME'])
wmata_stations = wmata_stations.set_crs('EPSG:4326')
wmata_stations = wmata_stations.to_crs('EPSG:2248')
wmata_stations.head(5)

correction_dict = {}
for berth_id in wmata_stoptimes['BERTH_ID'].unique():
    if pd.notna(berth_id):
        matches = wmata_stations[[berth_id in station_list for station_list in wmata_stations['STATION_ID_LIST'].values]]
        
        if len(matches) > 0:
            correction_dict[berth_id] = matches['STATION_ID'].values[0]

wmata_stoptimes['BERTH_ID'] = wmata_stoptimes['BERTH_ID'].replace(correction_dict)
wmata_stoptimes['NEXT_BERTH'] = wmata_stoptimes['NEXT_BERTH'].replace(correction_dict)
wmata_stoptimes = wmata_stoptimes.rename(columns={'BERTH_ID': 'STOP_ID',
                                                  'NEXT_BERTH': 'NEXT_STOP_ID'})
wmata_stoptimes.head(40)

### Save file

wmata_stoptimes.to_csv(f"Intermediate Results\\WMATA_stoptimes_{wmata_dataset_year}.csv", index=False)
